Provide proxies for I2C and SPI
